requirements.txt
aiogram>=3.25.0
google-genai>=1.63.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
